+++ SendBird.d.ts
@@ -1,5 +1,5 @@
 /**
- * Type Definitions for SendBird SDK v3.0.116
+ * Type Definitions for SendBird SDK v3.0.117
  * homepage: https://sendbird.com/
  * git: https://github.com/sendbird/SendBird-SDK-JavaScript
  */
@@ -11,7 +11,7 @@ declare const SendBird: SendBirdStatic;
 
 interface SendBirdStatic {
   version: number;
-  new({ appId }: { appId: string }): SendBird.SendBirdInstance;
+  new ({ appId }: { appId: string }): SendBird.SendBirdInstance;
   getInstance(): SendBird.SendBirdInstance;
 }
 
@@ -42,7 +42,7 @@ declare namespace SendBird {
 
   type pushTokens = {
     deviceTokens: Array<string>;
-    type: 'gcm' | 'apns' | 'apns_voip';
+    type: "gcm" | "apns" | "apns_voip";
     hasMore: boolean;
     token: string;
   };
@@ -109,8 +109,8 @@ declare namespace SendBird {
     createUserListQuery(): UserListQuery; // DEPRECATED
     createUserListQuery(userIds: Array<string>): UserListQuery; // DEPRECATED
     createApplicationUserListQuery(): ApplicationUserListQuery;
-
     createBlockedUserListQuery(): BlockedUserListQuery;
+    createMessageSearchQuery(keyword: String, options: MessageSearchQueryOptions): MessageSearchQuery;
 
     blockUser(userToBlock: User, callback?: userCallback): void;
     blockUserWithUserId(userToBlock: string, callback?: userCallback): void;
@@ -133,12 +133,12 @@ declare namespace SendBird {
     unregisterAPNSPushTokenAllForCurrentUser(callback?: commonCallback): void;
 
     unregisterPushTokenAllForCurrentUser(callback?: commonCallback): void; // This removes all push tokens including APNS/GCM
-    getMyPushTokensByToken(token: string, type: 'gcm' | 'apns' | 'apns_voip', callback: getMyPushTokensHandler): void;
+    getMyPushTokensByToken(token: string, type: "gcm" | "apns" | "apns_voip", callback: getMyPushTokensHandler): void;
 
     setPushTemplate(templateName: string, callback?: pushSettingCallback): void;
     getPushTemplate(callback?: pushSettingCallback): void;
 
-    setPushTriggerOption(pushTriggerOption: 'all' | 'mention_only' | 'off', callback: commonCallback): void;
+    setPushTriggerOption(pushTriggerOption: "all" | "mention_only" | "off", callback: commonCallback): void;
     getPushTriggerOption(callback: getPushTriggerOptionCallback): void;
 
     setDoNotDisturb(
@@ -184,10 +184,7 @@ declare namespace SendBird {
     getGroupChannelCount(memberStateFilter: string, callback: commonCallback): void;
 
     getUnreadItemCount(keys: Array<string>, callback: commonCallback): void;
-    getTotalUnreadMessageCount(
-      groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams,
-      callback: groupChannelCountCallback
-    ): void;
+    getTotalUnreadMessageCount(groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams, callback: groupChannelCountCallback): void;
     getTotalUnreadMessageCount(callback: groupChannelCountCallback): void;
     getTotalUnreadMessageCount(channelCustomTypes: Array<string>, callback: groupChannelCountCallback): void;
     getTotalUnreadChannelCount(callback: groupChannelCountCallback): void;
@@ -196,28 +193,10 @@ declare namespace SendBird {
     getSubscribedCustomTypeTotalUnreadMessageCount(): number;
     getSubscribedCustomTypeUnreadMessageCount(customType: string): number;
 
-    getMyGroupChannelChangeLogsByToken(
-      token: string,
-      customTypes: Array<string>,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
-    getMyGroupChannelChangeLogsByToken(
-      token: string,
-      customTypes: Array<string>,
-      includeEmpty: boolean,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
-    getMyGroupChannelChangeLogsByTimestamp(
-      ts: number,
-      customTypes: Array<string>,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
-    getMyGroupChannelChangeLogsByTimestamp(
-      ts: number,
-      customTypes: Array<string>,
-      includeEmpty: boolean,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
+    getMyGroupChannelChangeLogsByToken(token: string, customTypes: Array<string>, callback: getGroupChannelChangeLogsHandler): void;
+    getMyGroupChannelChangeLogsByToken(token: string, customTypes: Array<string>, includeEmpty: boolean, callback: getGroupChannelChangeLogsHandler): void;
+    getMyGroupChannelChangeLogsByTimestamp(ts: number, customTypes: Array<string>, callback: getGroupChannelChangeLogsHandler): void;
+    getMyGroupChannelChangeLogsByTimestamp(ts: number, customTypes: Array<string>, includeEmpty: boolean, callback: getGroupChannelChangeLogsHandler): void;
   }
 
   interface Options {
@@ -231,9 +210,24 @@ declare namespace SendBird {
     limit: number;
     next(callback: userListQueryCallback): void;
   }
+  interface MessageSearchQuery {
+    hasNext: boolean;
+    isLoading: boolean;
+    next(callback: messageListCallback): void;
+  }
+  interface MessageSearchQueryOptions {
+    limit: number;
+    reverse: boolean;
+    exactMatch: boolean;
+    channelUrl: string;
+    channelCustomType: string;
+    messageTimestampFrom: number;
+    messageTimestampTo: number;
+    order: "score" | "ts";
+  }
 
   interface UserEventHandlerStatic {
-    new(): UserEventHandler;
+    new (): UserEventHandler;
   }
   interface UserEventHandler {
     onFriendsDiscovered(users: Array<User>): void;
@@ -241,7 +235,7 @@ declare namespace SendBird {
   }
 
   interface ChannelHandlerStatic {
-    new(): ChannelHandler;
+    new (): ChannelHandler;
   }
   interface ChannelHandler {
     onMessageReceived(channel: OpenChannel | GroupChannel, message: AdminMessage | UserMessage | FileMessage): void;
@@ -275,7 +269,7 @@ declare namespace SendBird {
   }
 
   interface ConnectionHandlerStatic {
-    new(): ConnectionHandler;
+    new (): ConnectionHandler;
   }
   interface ConnectionHandler {
     onReconnectStarted(): void;
@@ -324,30 +318,30 @@ declare namespace SendBird {
   }
 
   interface GroupChannelTotalUnreadMessageCountParams {
-    new(): GroupChannelTotalUnreadMessageCountParams;
+    new (): GroupChannelTotalUnreadMessageCountParams;
     channelCustomTypesFilter: Array<string>;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
+    superChannelFilter: "all" | "super" | "nonsuper";
   }
 
   interface UserMessageParams {
-    new(): UserMessageParams;
+    new (): UserMessageParams;
     message: string;
     data: string;
     customType: string;
     targetLanguages: Array<string>; // DEPRECATED
     translationTargetLanguages: Array<string>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
     metaArrayKeys: Array<string>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
   }
   interface UserMessage extends BaseMessageInstance {
     message: string;
     sender: Sender;
     reqId: string;
     translations: Object;
-    requestState: 'none' | 'pending' | 'failed' | 'succeeded';
+    requestState: "none" | "pending" | "failed" | "succeeded";
     requestedMentionUserIds: Array<string>;
     errorCode: number;
     isResendable(): boolean;
@@ -357,7 +351,7 @@ declare namespace SendBird {
   }
 
   interface FileMessageParams {
-    new(): FileMessageParams;
+    new (): FileMessageParams;
     file: File;
     fileUrl: string;
     fileName: string;
@@ -366,11 +360,11 @@ declare namespace SendBird {
     data: string;
     customType: string;
     thumbnailSizes: Array<ThumbnailSize>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
     metaArrayKeys: Array<string>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
   }
   interface FileMessage extends BaseMessageInstance {
     sender: Sender;
@@ -380,7 +374,7 @@ declare namespace SendBird {
     size: number;
     type: string;
     thumbnails: Array<ThumbnailObject>;
-    requestState: 'none' | 'pending' | 'failed' | 'succeeded';
+    requestState: "none" | "pending" | "failed" | "succeeded";
     requestedMentionUserIds: Array<string>;
     errorCode: number;
     isResendable(): boolean;
@@ -437,7 +431,7 @@ declare namespace SendBird {
   }
 
   interface Member extends User {
-    state: 'invited' | 'joined';
+    state: "invited" | "joined";
     isBlockedByMe: boolean;
     isBlockingMe: boolean;
   }
@@ -489,21 +483,11 @@ declare namespace SendBird {
     getMessageChangeLogsByToken(callback: getMessageChangeLogsHandler): void;
     getMessageChangeLogsByToken(token: string, callback: getMessageChangeLogsHandler): void;
     getMessageChangeLogsByToken(token: string, includeMetaArray: boolean, callback: getMessageChangeLogsHandler): void;
-    getMessageChangeLogsByToken(
-      token: string,
-      includeMetaArray: boolean,
-      includeReaction: boolean,
-      callback: getMessageChangeLogsHandler
-    ): void;
+    getMessageChangeLogsByToken(token: string, includeMetaArray: boolean, includeReaction: boolean, callback: getMessageChangeLogsHandler): void;
 
     getMessageChangeLogsByTimestamp(ts: number, callback: getMessageChangeLogsHandler): void;
     getMessageChangeLogsByTimestamp(ts: number, includeMetaArray: boolean, callback: getMessageChangeLogsHandler): void;
-    getMessageChangeLogsByTimestamp(
-      ts: number,
-      includeMetaArray: boolean,
-      includeReaction: boolean,
-      callback: getMessageChangeLogsHandler
-    ): void;
+    getMessageChangeLogsByTimestamp(ts: number, includeMetaArray: boolean, includeReaction: boolean, callback: getMessageChangeLogsHandler): void;
 
     getMyMutedInfo(callback: getMyMutedInfoHandler): void;
     createOperatorListQuery(): OperatorListQuery;
@@ -769,30 +753,9 @@ declare namespace SendBird {
     sendFileMessage(file: File, callback: messageCallback): FileMessage;
     sendFileMessage(file: File, data: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(file: File, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      data: string,
-      customType: string,
-      thumbnailSizes: Array<ThumbnailSize>,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, customType: string, thumbnailSizes: Array<ThumbnailSize>, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, name: string, type: string, size: number, data: string, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, name: string, type: string, size: number, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: File,
       name: string,
@@ -804,25 +767,10 @@ declare namespace SendBird {
       callback: messageCallback
     ): FileMessage;
 
-    sendFileMessage(
-      fileMessageParams: FileMessageParams,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage;
+    sendFileMessage(fileMessageParams: FileMessageParams, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage;
     sendFileMessage(file: File, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage;
-    sendFileMessage(
-      file: File,
-      data: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      data: string,
-      customType: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, customType: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: File,
       data: string,
@@ -865,38 +813,12 @@ declare namespace SendBird {
     sendFileMessage(file: string, callback: messageCallback): FileMessage;
     sendFileMessage(file: string, data: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(file: string, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): FileMessage;
+    sendFileMessage(file: string, name: string, type: string, size: number, data: string, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, name: string, type: string, size: number, data: string, customType: string, callback: messageCallback): FileMessage;
 
     sendFileMessage(file: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      data: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      data: string,
-      customType: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, data: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, data: string, customType: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: string,
       name: string,
@@ -917,55 +839,32 @@ declare namespace SendBird {
       callback: messageCallback
     ): FileMessage; // DEPRECATED
 
-    sendFileMessages(
-      fileMessageParamsList: Array<FileMessageParams>,
-      callbackObject: fileMessagesCallbackObject
-    ): Array<FileMessage>;
+    sendFileMessages(fileMessageParamsList: Array<FileMessageParams>, callbackObject: fileMessagesCallbackObject): Array<FileMessage>;
 
     /** UserMessage  */
     sendUserMessage(userMessageParams: UserMessageParams, callback: messageCallback): UserMessage;
     sendUserMessage(message: string, callback: messageCallback): UserMessage;
     sendUserMessage(message: string, data: string, callback: messageCallback): UserMessage; // DEPRECATED
     sendUserMessage(message: string, data: string, customType: string, callback: messageCallback): UserMessage; // DEPRECATED
-    sendUserMessage(
-      message: string,
-      data: string,
-      customType: string,
-      translationTargetLanguages: Array<string>,
-      callback: messageCallback
-    ): UserMessage;
+    sendUserMessage(message: string, data: string, customType: string, translationTargetLanguages: Array<string>, callback: messageCallback): UserMessage;
 
     resendUserMessage(userMessage: UserMessage, callback: messageCallback): void;
     resendFileMessage(fileMessage: FileMessage, callback: messageCallback): void;
     resendFileMessage(fileMessage: FileMessage, file: Blob, callback: messageCallback): void;
 
-    translateUserMessage(
-      message: UserMessage,
-      translationTargetLanguages: Array<string>,
-      callback: messageCallback
-    ): void;
+    translateUserMessage(message: UserMessage, translationTargetLanguages: Array<string>, callback: messageCallback): void;
 
     /** Edit Message  */
     updateFileMessage(messageId: number, data: string, customType: string, callback: messageCallback): void;
     updateFileMessage(messageId: number, fileMessageParams: FileMessageParams, callback: messageCallback): void;
-    updateUserMessage(
-      messageId: number,
-      message: string,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): void;
+    updateUserMessage(messageId: number, message: string, data: string, customType: string, callback: messageCallback): void;
     updateUserMessage(messageId: number, userMessageParams: UserMessageParams, callback: messageCallback): void;
     deleteMessage(message: FileMessage | UserMessage, callback: commonCallback): void;
     cancelUploadingFileMessage(messageReqId: string, callback: cancelUploadingFileMessageCallback): boolean;
 
     /** Reaction */
     addReaction(message: UserMessage | FileMessage | AdminMessage, key: string, callback: reactionEventCallback): void;
-    deleteReaction(
-      message: UserMessage | FileMessage | AdminMessage,
-      key: string,
-      callback: reactionEventCallback
-    ): void;
+    deleteReaction(message: UserMessage | FileMessage | AdminMessage, key: string, callback: reactionEventCallback): void;
 
     /** MetaData */
     createMetaData(metaDataMap: Object, callback: commonCallback): void;
@@ -988,64 +887,26 @@ declare namespace SendBird {
     deleteAllMetaCounters(callback: commonCallback): void;
 
     /** MessageMetaArray */
-    createMessageMetaArrayKeys(
-      message: UserMessage | FileMessage | AdminMessage,
-      keys: Array<string>,
-      callback: commonCallback
-    ): void;
-    deleteMessageMetaArrayKeys(
-      message: UserMessage | FileMessage | AdminMessage,
-      keys: Array<string>,
-      callback: commonCallback
-    ): void;
-    addMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Array<MessageMetaArray>,
-      callback: commonCallback
-    ): void;
-    removeMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Array<MessageMetaArray>,
-      callback: commonCallback
-    ): void;
+    createMessageMetaArrayKeys(message: UserMessage | FileMessage | AdminMessage, keys: Array<string>, callback: commonCallback): void;
+    deleteMessageMetaArrayKeys(message: UserMessage | FileMessage | AdminMessage, keys: Array<string>, callback: commonCallback): void;
+    addMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Array<MessageMetaArray>, callback: commonCallback): void;
+    removeMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Array<MessageMetaArray>, callback: commonCallback): void;
 
-    addMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Object,
-      callback: commonCallback
-    ): void; // DEPRECATED
-    removeMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Object,
-      callback: commonCallback
-    ): void; // DEPRECATED
+    addMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Object, callback: commonCallback): void; // DEPRECATED
+    removeMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Object, callback: commonCallback): void; // DEPRECATED
 
     /** Report */
     report(category: string, description: string, callback: commonCallback): void;
     reportUser(user: User, category: string, description: string, callback: commonCallback): void;
-    reportMessage(
-      message: UserMessage | FileMessage | AdminMessage,
-      category: string,
-      description: string,
-      callback: commonCallback
-    ): void;
+    reportMessage(message: UserMessage | FileMessage | AdminMessage, category: string, description: string, callback: commonCallback): void;
   }
 
-  type messageListCallback = (
-    messageList: Array<UserMessage | FileMessage | AdminMessage>,
-    error: SendBirdError
-  ) => void;
+  type messageListCallback = (messageList: Array<UserMessage | FileMessage | AdminMessage>, error: SendBirdError) => void;
   interface MessageListQuery {
     // DEPRECATED
     next(messageTimestamp: number, limit: number, reverse: boolean, callback: messageListCallback): void;
     prev(messageTimestamp: number, limit: number, reverse: boolean, callback: messageListCallback): void;
-    load(
-      messageTimestamp: number,
-      prevLimit: number,
-      nextLimit: number,
-      reverse: boolean,
-      callback: messageListCallback
-    ): void;
+    load(messageTimestamp: number, prevLimit: number, nextLimit: number, reverse: boolean, callback: messageListCallback): void;
   }
 
   interface PreviousMessageListQuery {
@@ -1082,13 +943,7 @@ declare namespace SendBird {
     createBannedUserListQuery(): BannedUserListQuery;
 
     updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: openChannelCallback): void;
-    updateChannel(
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      operatorUserIds: Array<string> | string,
-      callback: openChannelCallback
-    ): void; // DEPRECATED
+    updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void; // DEPRECATED
     updateChannel(
       name: string,
       coverUrlOrImageFile: string | File,
@@ -1203,13 +1058,7 @@ declare namespace SendBird {
 
     createChannel(callback: openChannelCallback): void;
     createChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: openChannelCallback): void; // DEPRECATED
-    createChannel(
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      operatorUserIds: Array<string> | string,
-      callback: openChannelCallback
-    ): void; // DEPRECATED
+    createChannel(name: string, coverUrlOrImageFile: string | File, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void; // DEPRECATED
     createChannel(
       name: string,
       coverUrlOrImageFile: string | File,
@@ -1256,7 +1105,7 @@ declare namespace SendBird {
     scheduledId: number;
     scheduledDateTimeString: string;
     scheduledTimezone: string;
-    status: 'scheduled' | 'sent' | 'canceled' | 'failed';
+    status: "scheduled" | "sent" | "canceled" | "failed";
     createdAt: number;
     updatedAt: number;
     channelUrl: string;
@@ -1266,9 +1115,9 @@ declare namespace SendBird {
     data: string;
     metaArray: Object; // DEPRECATED
     metaArrays: Array<MessageMetaArray>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUsers: Array<User>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     translationTargetLanguages: Array<string>;
     errorMessage: string;
     errorCode: number;
@@ -1283,7 +1132,7 @@ declare namespace SendBird {
    * GroupChannel
    */
   interface GroupChannelParams {
-    new(): GroupChannelParams;
+    new (): GroupChannelParams;
     isDistinct: boolean;
     isSuper: boolean;
     isPublic: boolean;
@@ -1307,16 +1156,16 @@ declare namespace SendBird {
   }
 
   interface ScheduledUserMessageParams {
-    new(): ScheduledUserMessageParams;
+    new (): ScheduledUserMessageParams;
     message: string;
     data: string;
     customType: string;
     translationTargetLanguages: Array<string>;
     metaArrayKeys: Array<string>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     year: number;
     month: number;
     day: number;
@@ -1333,39 +1182,36 @@ declare namespace SendBird {
   }
 
   interface MessageMetaArray {
-    new(key: string, value: Array<string>): MessageMetaArray;
+    new (key: string, value: Array<string>): MessageMetaArray;
     key: string;
     value: Array<string>;
   }
   interface Reaction {
-    new(): Reaction;
+    new (): Reaction;
     key: string;
     userIds: Array<string>;
     updatedAt: number;
   }
   interface ReactionEvent {
-    new(): ReactionEvent;
+    new (): ReactionEvent;
     messageId: string;
     userId: string;
     key: string;
-    operation: 'add' | 'delete';
+    operation: "add" | "delete";
     updatedAt: number;
   }
 
   type groupChannelCallback = (groupChannel: GroupChannel, error: SendBirdError) => void;
   type distinctGroupChannelCallback = (response: DistinctGroupChannelResponse, error: SendBirdError) => void;
   type getPushPreferenceCallback = (isPushOn: boolean, error: SendBirdError) => void;
-  type getPushTriggerOptionCallback = (
-    pushTriggerOption: 'all' | 'mention_only' | 'off' | 'default',
-    error: SendBirdError
-  ) => void;
+  type getPushTriggerOptionCallback = (pushTriggerOption: "all" | "mention_only" | "off" | "default", error: SendBirdError) => void;
   interface GroupChannel extends BaseChannel {
     isHidden: boolean;
     isDistinct: boolean;
     isSuper: boolean;
     isPublic: boolean;
     isPushEnabled: boolean; // DEPRECATED
-    myPushTriggerOption: 'default' | 'all' | 'mention_only' | 'off';
+    myPushTriggerOption: "default" | "all" | "mention_only" | "off";
     myCountPreference: string;
     lastMessage: UserMessage | FileMessage | AdminMessage;
     unreadMessageCount: number;
@@ -1373,13 +1219,13 @@ declare namespace SendBird {
     members: Array<Member>;
     memberCount: number;
     joinedMemberCount: number;
-    myMemberState: 'none' | 'joined' | 'invited';
-    myRole: 'operator' | 'none';
-    myMutedState: 'muted' | 'unmuted';
+    myMemberState: "none" | "joined" | "invited";
+    myRole: "operator" | "none";
+    myMutedState: "muted" | "unmuted";
     inviter: User;
     invitedAt: number;
     isAccessCodeRequired: boolean;
-    hiddenState: 'unhidden' | 'hidden_allow_auto_unhide' | 'hidden_prevent_auto_unhide';
+    hiddenState: "unhidden" | "hidden_allow_auto_unhide" | "hidden_prevent_auto_unhide";
     isDiscoverable: boolean;
     myLastRead: number;
     messageOffsetTimestamp: number;
@@ -1391,13 +1237,7 @@ declare namespace SendBird {
 
     updateChannel(groupChannelParams: GroupChannelParams, callback: groupChannelCallback): void;
     updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void;
-    updateChannel(
-      isDistinct: boolean,
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      callback: groupChannelCallback
-    ): void; // DEPRECATED
+    updateChannel(isDistinct: boolean, name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void; // DEPRECATED
     updateChannel(
       isDistinct: boolean,
       name: string,
@@ -1440,15 +1280,9 @@ declare namespace SendBird {
 
     setPushPreference(pushOn: boolean, callback: commonCallback): void; // DEPRECATED
     getPushPreference(callback: getPushPreferenceCallback): void; // DEPRECATED
-    setMyPushTriggerOption(
-      pushTriggerOption: 'all' | 'mention_only' | 'off' | 'default',
-      callback: getPushTriggerOptionCallback
-    ): void;
+    setMyPushTriggerOption(pushTriggerOption: "all" | "mention_only" | "off" | "default", callback: getPushTriggerOptionCallback): void;
     getMyPushTriggerOption(callback: getPushTriggerOptionCallback): void;
-    setMyCountPreference(
-      preference: 'all' | 'unread_message_count_only' | 'unread_mention_count_only' | 'off',
-      callback: commonCallback
-    ): void;
+    setMyCountPreference(preference: "all" | "unread_message_count_only" | "unread_mention_count_only" | "off", callback: commonCallback): void;
 
     createMemberListQuery(): GroupChannelMemberListQuery;
     createBannedUserListQuery(): BannedUserListQuery;
@@ -1468,10 +1302,7 @@ declare namespace SendBird {
     freeze(callback: commonCallback): void;
     unfreeze(callback: commonCallback): void;
 
-    registerScheduledUserMessage(
-      scheduledUserMessageParams: ScheduledUserMessageParams,
-      callback: scheduledUserMessageCallback
-    ): void;
+    registerScheduledUserMessage(scheduledUserMessageParams: ScheduledUserMessageParams, callback: scheduledUserMessageCallback): void;
   }
 
   type groupChannelCountCallback = (count: number, error: SendBirdError) => void;
@@ -1482,10 +1313,7 @@ declare namespace SendBird {
     createPublicGroupChannelListQuery(): PublicGroupChannelListQuery;
 
     getUnreadItemCount(keys: Array<string>, callback: commonCallback): void; // DEPRECATED
-    getTotalUnreadMessageCount(
-      groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams,
-      callback: groupChannelCountCallback
-    ): void; // DEPRECATED
+    getTotalUnreadMessageCount(groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams, callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadMessageCount(callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadMessageCount(channelCustomTypes: Array<string>, callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadChannelCount(callback: groupChannelCountCallback): void; // DEPRECATED
@@ -1511,19 +1339,11 @@ declare namespace SendBird {
       customType: string,
       callback: groupChannelCallback
     ): void;
-    createDistinctChannelIfNotExist(
-      groupChannelParams: GroupChannelParams,
-      callback: distinctGroupChannelCallback
-    ): void;
+    createDistinctChannelIfNotExist(groupChannelParams: GroupChannelParams, callback: distinctGroupChannelCallback): void;
 
     createChannelWithUserIds(userIds: Array<string>, callback: groupChannelCallback): void;
     createChannelWithUserIds(userIds: Array<string>, isDistinct: boolean, callback: groupChannelCallback): void; // DEPRECATED
-    createChannelWithUserIds(
-      userIds: Array<string>,
-      isDistinct: boolean,
-      customType: string,
-      callback: groupChannelCallback
-    ): void; // DEPRECATED
+    createChannelWithUserIds(userIds: Array<string>, isDistinct: boolean, customType: string, callback: groupChannelCallback): void; // DEPRECATED
     createChannelWithUserIds(
       userIds: Array<string>,
       isDistinct: boolean,
@@ -1553,9 +1373,9 @@ declare namespace SendBird {
     limit: number;
     hasNext: boolean;
     isLoading: boolean;
-    mutedMemberFilter: 'all' | 'muted' | 'unmuted';
-    operatorFilter: 'all' | 'operator' | 'nonoperator'; // DEPRECATED
-    memberStateFilter: 'all' | 'joined_only' | 'invited_only' | 'invited_by_friend' | 'invited_by_non_friend';
+    mutedMemberFilter: "all" | "muted" | "unmuted";
+    operatorFilter: "all" | "operator" | "nonoperator"; // DEPRECATED
+    memberStateFilter: "all" | "joined_only" | "invited_only" | "invited_by_friend" | "invited_by_non_friend";
     nicknameStartsWithFilter: string;
 
     next(callback: groupChannelMemberListQueryCallback): void;
@@ -1570,25 +1390,25 @@ declare namespace SendBird {
     hasNext: boolean;
     isLoading: boolean;
     includeEmpty: boolean;
-    order: 'latest_last_message' | 'chronological' | 'channel_name_alphabetical' | 'metadata_value_alphabetical';
+    order: "latest_last_message" | "chronological" | "channel_name_alphabetical" | "metadata_value_alphabetical";
     userIdsFilter: Array<string>; // DEPRECATED
     userIdsFilterExactMatch: boolean; // DEPRECATED
-    queryType: 'AND' | 'OR'; // DEPRECATED
+    queryType: "AND" | "OR"; // DEPRECATED
     userIdsExactFilter: Array<string>;
     userIdsIncludeFilter: Array<string>;
-    userIdsIncludeFilterQueryType: 'AND' | 'OR';
+    userIdsIncludeFilterQueryType: "AND" | "OR";
     nicknameContainsFilter: string;
     channelNameContainsFilter: string;
     customTypeFilter: string; // DEPRECATED
     customTypesFilter: Array<string>;
     customTypeStartsWithFilter: string;
     channelUrlsFilter: Array<string>;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
-    publicChannelFilter: 'all' | 'public' | 'private';
+    superChannelFilter: "all" | "super" | "nonsuper";
+    publicChannelFilter: "all" | "public" | "private";
     metadataOrderKeyFilter: string;
-    memberStateFilter: 'all' | 'joined_only' | 'invited_only' | 'invited_by_friend' | 'invited_by_non_friend';
-    hiddenChannelFilter: 'unhidden_only' | 'hidden_only' | 'hidden_allow_auto_unhide' | 'hidden_prevent_auto_unhide';
-    unreadChannelFilter: 'all' | 'unread_message';
+    memberStateFilter: "all" | "joined_only" | "invited_only" | "invited_by_friend" | "invited_by_non_friend";
+    hiddenChannelFilter: "unhidden_only" | "hidden_only" | "hidden_allow_auto_unhide" | "hidden_prevent_auto_unhide";
+    unreadChannelFilter: "all" | "unread_message";
 
     setSearchFilter(fields: Array<string>, queryString: string): void;
     serialize(): object;
@@ -1600,13 +1420,13 @@ declare namespace SendBird {
     hasNext: boolean;
     isLoading: boolean;
     includeEmpty: boolean;
-    order: 'chronological' | 'channel_name_alphabetical' | 'metadata_value_alphabetical';
+    order: "chronological" | "channel_name_alphabetical" | "metadata_value_alphabetical";
     channelNameContainsFilter: string;
     channelUrlsFilter: Array<string>;
     customTypesFilter: Array<string>;
     customTypeStartsWithFilter: string;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
-    membershipFilter: 'all' | 'joined';
+    superChannelFilter: "all" | "super" | "nonsuper";
+    membershipFilter: "all" | "joined";
     metadataOrderKeyFilter: string;
     next(callback: groupChannelListQueryCallback): void;
   }